wattwise/Flask API + MySQL DB/requirements.txt
Flask
gunicorn
python-dotenv
mysql-connector-python
pandas
